Add some comments to make the code easier to follow
